Ability to load datasets from text file
Defining new datasets in the code seemed a bit cumbersome so I made the code instead read in the datasets from an external text file that is formatted in a specific way. The text file also contains instructions on how to define new datasets. It also contains all of Joe's original datasets, which have been deleted from the code file.