Initial restructuring for improved revisiting
The code has been restructured such that class_stats now returns the confusion matrix, as opposed to the calculated statistics.
The idea being that the the results can then be reproduced from a saved confusion matrix, if the outputs and targets have been lost.

However, I am now thinking that this is not a good decision. stats_class is the all encapsulating function because it intends to do everything and return the statistics. If a specific use is required, then that should be made into a separate function that just encapsulates the specific use case. I.e. if I just wanted to collect the confusion matrix at the current time so that the statistics can possibly be calculated and displayed at a later time, then stats_class should not be used to do this with a bunch of parameters saying whether or not to display and plot and having to give data names etc. Instead, construct_confusion_matrix (with the only parameters being the outputs and targets) should be used, and a new function calculate_and_display_stats (with the parameters being the confusion matrix, data name, and some booleans for display options) should be called later on when desired. calculate_and_display_stats would only serve to encapsulate the calculate_stats and display_stats functions into a single function.

WARNING: The code does not yet fully reflect the desired changes as mentioned above. The code has not been tested or re-documented. This commit is purely for saving my progress across public computers whilst my laptop is dead.