Add support for multi-modal classification data
data_presets.txt now has an optional extra parameter for defining labels for each distribution. Multiple distributions using the same label, will belong to the same class.

The new parameter is optional. If the parameter isn't entered, then each distribution will have it's own class - just like before.

The plot now also supports such labels. The label and the target will be used in the plot. If the label and the target happen to be the same, or the labels are not provided, then just the target will be used - as before.