Basic save hyperparameters and results to .txt
Additionally, turned off plots.

At the moment, nothing is written until the entire test is complete. This means that crashes will cause all data to be lost from the run.

Furthermore, at the moment, there is not a easy functionality for continuing off the test from the last number of patterns saved.

The code now is very rudimentary and basic, but is a start pointing at least.