Better text file writing
The text file that contains the hyperparameters and the optimal lambda paired with the hyperparameter value (currently only number of patterns) is now more compact.

Each text file is also generated with a basically unique hexadecimal ID number which depends on the hyperparameters: learning rate, number of epochs, the lambdas we test, how much validation data we have, how many hidden nodes we have in the layer and the name of the dataset. Changing any of these results in a completely different ID, but the process used to calculate the ID should be deterministic and consistent across all machines.

The fixed seed is also included in the text file (and its title) but does not influence the ID, so if many runs with the same hyperparameters are conducted they will have the same ID except the final digit(s). This should allow us to efficiently loop over all the text files to collect the different optimal lambdas and calculate the mean etc.

Also added some exceptions if you try to run the code with seed = -1 or stochastic gradient descent.

To do:
- Read the contents of the text file in such a way that we can calculate the mean and SEM for each data point and plot it.
- Update optimize_hidden.py with all this garbage.
- Maybe disable some stuff, such as calculating validation loss ever epoch, to speed up computational time.
- Come up with good distributions for the lambdas and hyperparameter values we test
- Make some banger plots