Automatic text file reading
errorbars.py can now read the results (best lambda for a given number of patterns) from multiple text files that share the same id (which means all hyperparameters are identical) but where ran with different seeds. It is not necessary for each number of pattern to have the same amount of best lambda results and there is no need to specify precisely which numbers of patterns that have been tested, the program automatically figures this out and the functions that calculate the mean and SEM work on an arbitrary array size. This means that once you have your text files all you have to do is tell the program what seeds you used and what the id is, it will then generate a plot with error bars based on the result.

The only change in optimize_patterns.py is that it throws an exception if you try to create a text file with a name that already exists. This is to avoid complications where errorbars.py thinks that data generated from the same seed is actually new independent results.

To do:
- Update optimize_hidden.py with all this garbage and add a toggle in errorbars.py for the hyperparameter you want to vary
- Maybe disable some stuff, such as calculating validation loss ever epoch, to speed up computational time.
- Come up with good distributions for the lambdas and hyperparameter values we test
- Make some banger plots